Documentation for GeneralSettingsClass now uses numpy-style dosctrings.
