+++ mirp/settings/settingsClass.py
@@ -4,71 +4,7 @@ import numpy as np
 
 from typing import Union, List, Tuple, Any, Optional, Iterable
 
-
-class GeneralSettingsClass:
-
-    def __init__(
-            self,
-            by_slice: Union[str, bool] = False,
-            config_str: str = "",
-            divide_disconnected_roi: str = "keep_as_is",
-            no_approximation: bool = False,
-            **kwargs):
-        """
-        Sets general experiment parameters.
-
-        :param by_slice: Defines whether calculations should be performed in 2D (True) or 3D (False),
-            or alternatively only in the largest slice (Largest). Default: False
-        :param config_str: Sets a configuration string, which can be used to differentiate files or analyses created
-            from the same data but a different configuration.
-        :param divide_disconnected_roi: Defines how ROI segmentations are treated after being loaded. Can be
-            "keep_as_is" (default), "keep_largest", "combine".
-        :param no_approximation: Disables approximation of features, such as Geary's c-measure. Can be True or False
-            (default).
-        :param kwargs: Unused keyword arguments.
-
-        :returns: A :class:`mirp.importSettings.GeneralSettingsClass` object with configured parameters.
-        """
-
-        # Parse and check slice information.
-        if isinstance(by_slice, str):
-            if by_slice.lower() in ["true", "t", "1"]:
-                by_slice = True
-                select_slice = "all"
-            elif by_slice.lower() in ["false", "f", "0"]:
-                by_slice = False
-                select_slice = "all"
-            elif by_slice.lower() in ["largest"]:
-                by_slice = True
-                select_slice = "largest"
-            else:
-                raise ValueError(
-                    f"The by_slice parameter should be true, false, t, f, 1, 0 or largest. Found: {by_slice}")
-
-        elif isinstance(by_slice, bool):
-            select_slice = "all"
-
-        else:
-            raise ValueError("The by_slice parameter should be a string or boolean.")
-
-        # Set by_slice and select_slice parameters.
-        self.by_slice: bool = by_slice
-        self.select_slice: str = select_slice
-
-        # Set configuration string.
-        self.config_str: str = config_str
-
-        # Check divide_disconnected_roi
-        if divide_disconnected_roi not in ["keep_as_is", "keep_largest", "combine"]:
-            raise ValueError(
-                f"The divide_disconnected_roi parameter should be 'keep_as_is', 'keep_largest', "
-                f"'combine'. Found: {divide_disconnected_roi}")
-
-        # Set divide_disconnected_roi
-        self.divide_disconnected_roi: str = divide_disconnected_roi
-
-        # Set approximation of features.
-        self.no_approximation: bool = no_approximation
+from mirp.settings.generalSettingsClass import GeneralSettingsClass
 
 
 class ImageInterpolationSettingsClass:
