+++ test/ibsi_2_response_map_test.py
@@ -5,9 +5,10 @@ import numpy as np
 import os
 
 from mirp.extractFeaturesAndImages import extract_images
-from mirp.settings.settingsClass import SettingsClass, GeneralSettingsClass, ImagePostProcessingClass,\
+from mirp.settings.settingsClass import SettingsClass, ImagePostProcessingClass,\
     ImageInterpolationSettingsClass, RoiInterpolationSettingsClass, ResegmentationSettingsClass,\
     ImagePerturbationSettingsClass, ImageTransformationSettingsClass, FeatureExtractionSettingsClass
+from mirp.settings.generalSettingsClass import GeneralSettingsClass
 
 # Find path to the test directory. This is because we need to read datafiles stored in subdirectories.
 CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
